requirements.txt
Flask==2.2.5
pymongo==4.3.3
dnspython==2.3.0
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==20.1.0
